+++ xterm256_colors/__init__.py
@@ -1,4 +1,4 @@
-__version__ = '0.1.3'
+__version__ = '0.1.4'
 
 from .colors import Back256, Fore256, XtermCodes, XtermColor
 from .utils import (